+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.2.1]
+### Changed
+- Bump Microsoft.Identity.Web from 4.10.0 to 4.13.0
+- Bump OpenTelemetry.Exporter.OpenTelemetryProtocol from 1.15.3 to 1.16.0
+- Bump OpenTelemetry.Instrumentation.Http from 1.15.1 to 1.16.0
+- Bump OpenTelemetry.Instrumentation.AspNetCore from 1.15.2 to 1.16.0
+- OpenTelemetry.Instrumentation.Runtime from 1.15.1 to 1.16.0
+
 ## [1.2.0]
 ### Added
 - `BifrostAuthorizationMode` enum (`NoAuth`, `ServicePrincipal`, `SystemAssignedIdentity`, `UserAssignedIdentity`) to select the credential flow used by the exporter.